Create 1.19 LAB: Input and formatted output: House real estate summary
Sites like Zillow get input about house prices from a database and provide nice summaries for readers. Write a program with two inputs, current price and last month's price (both integers). Then, output a summary listing the price, the change since last month, and the estimated monthly mortgage computed as (currentPrice * 0.045) / 12 (Note: Output directly. Do not store in a variable.).

Ex: If the input is 200000 210000, the output is:

This house is $200000. The change is $-10000 since last month. 
The estimated monthly mortgage is $750.
Note: Getting the precise spacing, punctuation, and newlines exactly right is a key point of this assignment. Such precision is an important part of programming.