Create 2.24 LAB: Chemistry formula
People commonly write simple programs (even if not a CS major) to do a complex calculation. Below is the van der Waals equation for a real gas, relating Pressure, Volume, Temperature, for n moles of a gas:

The term in the left parentheses is known as the "ideal gas pressure", and in the right parentheses as the "ideal gas volume". R is known as the ideal gas constant and equals 8.314. (See https://en.wikipedia.org/wiki/VanderWaals_equation for more info).

Write a program that takes P, V, and n as inputs, and outputs T, for carbon dioxide for which a = 3.59 and b = 0.0427. You'll of course first need to rewrite the equation to solve for T, using simple algebra; don't try to simplify the equation though. All variables should be float. (Don't try to convert units. Assume inputs are positive.).

If the input is: 100 10 1
the output is:

119.808
TAs may examine your code closely for style (and some course points are for style). Make sure your code uses good variable names, following the class' style. It's a good idea to avoid writing the entire formula as one big equation. Use comments to explain things as appropriate. There are many ways to write even this simple program, so we would expect your program will be somewhat different from folks you might be getting help from or helping.

Hint: Try using temporary variables to hold a partial calculation.