Create 3.14 LAB: Largest number
Write a program whose inputs are three integers, and whose output is the largest of the three integers, followed by a count of the number of times that largest integer appears.

Ex: If the input is 7 15 3, the output is:

15
1

Ex: If the input is 9 9 8, the output is:

9
2
Hint: Start by writing code that just outputs the largest integer and submit to get most of the points. Then, update your code to output both the largest integer and the count. Incremental development is a great habit to practice!