Create 4.9 Lab: Random Animal
Part 1: Complete the survey.

Part 2: Peek-a-boo is a fun game that little kids like to play. To simulate this game on the computer, write a program that will generate a random number between 1 and 4. Then, will print to the screen the animal name associated to that number. The animal names used will be:

pig when a 1 is generated

cow when a 2 is generated

chicken when a 3 is generated

horse when a 4 is generated

If your program generates a 3, the output will be:

chicken

The player will then enter a 1 if they would like to play again or anything else to exit the program.

If the player enters "1 1 1 0", the output will be: horse chicken cow horse

For coding simplicity, follow each output animal by a space, even the last one.

Hint: To make testing easier, seed your random number generator with 0. Also, try to execute one iteration of the game before you repeat the game.