Create 5.13 LAB: Count input length without end-of-sentence punctuation
Given a line of text as input: (1) output the number of characters excluding the three characters commonly used for end-of-sentence punctuation( period, exclamation point, and question mark), (2) then output the number of end-of-sentence punctuation characters that were found. You can just do (1) to pass the first few test cases for partial credit, then do (2) for full credit.

Ex: If the input is "Listen, Sam! Calm down. Please.", the output is:

28
3
Ex: If the input is "What time is it? Time to get a watch! O.K., bye now.", the output is:

47
5
(because the periods in O.K. are counted, even though they don't actually end a sentence).