Create 5.16 LAB: Checker for integer string
Part 1: Complete the survey.

Part 2: Forms often allow a user to enter an integer. Write a program that takes in a string representing an integer as input, and outputs yes if every character is a digit 0-9.

Ex: If the input is 1995, the output is:

yes
Ex: If the input is 42,000 or 1995!, the output is:

no
Hint: Use a loop and the isdigit() function (don't forget to include the cctype library). Do not use break!