Create 5.18 LAB: Remove spaces
Write a program that removes all spaces from the given input. Use getline to get input from the user. Hint: Use two strings. One to hold the input and one to hold the the string with spaces removed.

Ex: If the input is "Hello my name is John." (excluding the quotes), the output is:

HellomynameisJohn.