Create 7.13 LAB: Count short words in a file
Write a program that gets a number that represents the largest number of characters a string can have to be considered a "short" word and also gets a file name as input. The program then reads words from a file and calls a function that will determine if the word is considered a "short" word. The program outputs the number of "short" words in that file. If the input is textdoc0.txt and that file contains:

Hi there cats and dogs. 
You're #  1 in my book!
then the output is:

6
(Short words were: Hi, and, #, 1, in, my)

Remember to do your error checks. Make sure the file opened and all the words are read. Then, close the file.

Your program must define and call a function IsShortWord that takes in a string and an int and returns a bool.