Create 8.8 LAB: Remove spaces - functions
Write a program that removes all spaces from the given input.

Ex: If the input is: "Hello my name is John." the output is:

HellomynameisJohn.
Your program must define and call the following function. The function should return a string representing the input string without spaces.
void RemoveSpaces(string &userString)

Note: This is a lab from a previous chapter that now requires the use of a function.